requirements.txt
Flask==3.0.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
Flask-Login==0.6.*
Flask-WTF==1.2.*
WTForms==3.1.*
email-validator==2.2.*
python-dotenv==1.0.*
werkzeug>=3.0
psycopg2-binary==2.9.*
requests==2.31.*
pytest==8.0.*
gunicorn==21.2.*

